:memo:  chore: make searchBar folder and move related files to it
